Implemented thumbnail overlays.

+++ YouTubeScrap/Data/Renderers/ThumbnailOverlayResumePlaybackRenderer.cs
@@ -5,6 +5,6 @@ namespace YouTubeScrap.Data.Renderers
     public class ThumbnailOverlayResumePlaybackRenderer
     {
         [JsonProperty("percentDurationWatched")]
-        public int PercentDurationWatched { get; set; }
+        public double PercentDurationWatched { get; set; }
     }
 }
